chore: rename `index_builders` to `indexBuilders`
